+++ RetroFE/Source/Video/GStreamerVideo.h
@@ -41,7 +41,7 @@ public:
     int getHeight();
     int getWidth();
     bool isPlaying();
-	void setVolume(double volume);
+	void setVolume(float volume);
 
 private:
     static void processNewBuffer (GstElement *fakesink, GstBuffer *buf, GstPad *pad, gpointer data);
@@ -63,7 +63,7 @@ private:
     int playCount_;
     std::string currentFile_;
     int numLoops_;
-	double volume_;
+	float volume_;
     double currentVolume_;
     int monitor_;
 };
